requirements.txt
kivy
requests
python-dotenv
